Crate oak_proxy_attestation_bin

Proxy Attestation Service implementation.

Proxy Attestation Service works as a Certificate Authority for TEE applications. It attests applications and creates signed certificates for TEE applications, which they use to create TLS connections with clients.

The work cycle of the proxy is represented in 2 stages.

First stage corresponds to the backend server attestation:

TEE application connects to the proxy

Proxy attests the application using a corresponding remote attestation protocol

Application sends the proxy a certificate signing request

Proxy creates a signed certificate and sends it back to the application

The certificate contains application’s TEE measurements

Second stage corresponds to the client connection:

Client remotely attests the proxy using a corresponding attestation protocol

Proxy sends the client its root certificate

Client trusts that this is a correct certificate, since it was sent via a secure channel created during the attestation process

Client connects to the application using TLS, and the application uses a TLS certificate previously signed by the proxy

Client checks that the certificate was signed by the root certificate and also checks TEE measurements

If all checks were successful, client establishes a secure connection

Note: Current version of Remote Attestation Service doesn’t have remote attestation.

Structs

Opt

Proxy

Functions

main 🔒